refactor : move section domain to line package
